Rename logger.js to backend_logger.js.
