+++ main.js
@@ -5,6 +5,6 @@ var connect = require('./connect');
 
 var app = newApp(connect.connection);
 
-var logger = require('./logger')();
+var logger = require('./backend_logger')();
 
 app.listen(process.env.PORT || 3000, function() {logger.info('Server is started');});
